Move editor class accessor to peas_base
